Add files via upload
CHANGE: Moved away from select alias to verbose select-object for consistency

CHANGE: When creating Problems as the default new Work Item, the Sender is marked in the Created By relationship instead of being skipped entirely

BUG: When creating a new Service Request based on a template if the template has activities, those activities are missing their Work Item prefix. This is a known issue with the SCSM SDK and has been addressed using Morten Meisler's Set-SCSMTemplateWithActivities (http://blog.ctglobalservices.com/service-manager-scsm/mme/set-scsmtemplatewithactivities-powershell-script/)

CHANGE: Modified Set-SCSMTemplateWithActivities to support Cireson SMA Connector Runbooks and Cireson PowerShell Activity should those activities be featured in templates

CHANGE: Instead of a new SR being in a "Submitted" status it now is placed in "New" status to honor workflows that may trigger on New to Submitted or New to In-Progress status changes

FEATURE: Added ability to create Change Requests as new default work item type, this now leverages the recently added Set-SCSMTemplateWithActivities

NOTE: With ability to create Change Requests as the new default Work Item, deployments may wish to mark the Sender as the Affected User on Change Requests. If this line is uncommented, it enables Cireson Portal customers to view Change Requests in the "My Requests" tab alongside Incidents and Service Requests. Furthermore, a vanilla SCSM deployment does not expose this relationship on CR forms nor does the Cireson Portal Form for Change Requests. If you wish to see this relationship on those forms you would have to either customize the form management pack through the SCSM Authoring tool and/or add the relationship to the ChangeRequest.js file for the Cireson portal.