Converting Invoke-WebRequest to Invoke-RestMethod (#94)
* Move WebRequest to RestMethod

Converted use of Invoke-WebRequest to Invoke-RestMethod for Cireson portal actions when Windows Authentication is not being used

* Convert API token into function

The PowerShell required to retrieve a Cireson API Token was moved into it's own dedicated function to when Windows Authentication is not used. Also caught an error with an extra "/" in the URL.

* Removing duplicate variables on Cireson calls

With the simplification of retrieving Cireson API tokens introduced, there are several Cireson based functions that have repetitive variables that could be condensed into a single one.

* Standardizing Cireson calls

This is more of a preference/consistency thing above all else in that all Cireson functions now declare their API URL specific variable before each Windows Auth if logic test. In this way they are more consistent and readable.