Simplify New-WorkItem by introducing Suggestion functions (#100)
* Condense New-WorkItem

Initial commit that removes some of the repetition seen in New-WorkItem when leveraging suggestions with/without Azure Cognitive Services. Creates two new functions Get-CiresonSuggestionURL and Send-CiresonSuggestionEmail that shorten New-WorkItem and simplify the logic when optionally leveraging Azure Cognitive Services

* Changing Send-CiresonSuggestionEmail

When Send-CiresonSuggestionEmail is called it was expecting an SCSM User Object, but when used it's missing the needed EmailAddress property as Email Addresses are technically a relationship in SCSM as opposed to a plain text field within Active Directory. Parameters have been changed, it's 6 references, and instead of passing the $affectedUser object instead am making use of the $from originally defined at the top of New-WorkItem