Initial Commit
This marks the first commit to the SettingsMP branch that contains the Visual Studio solution for building the Settings MP. It contains a single Solution with 2 projects that build the DLL UI, MP, and then the final MPB. It most closely mirrors v1.4.3 of the connector but is in an unfinished state. While the solution builds and can be be deployed into SCSM, the UI is missing several items as they relate to parameter/variable validation. Microsoft SCSM 2016 DLLs and MPs have been bundled with this so as to speed up everyone's initial build times as opposed to having to hunt down and resolve specific environment dependencies.