Upgrade to v1.4.6 (#102)
* Enforce template name (#99)

Updating the Get-SCSMObjectTemplate section to ensure only a single template object is pulled

* Change Incident Status on reply (#101)

* Change Incident Status on reply

Addressing feature request that changes Status on an Incident based on who actually did the reply. This introduces a new configuration variable, 3 "status to change to" style variables, and eventually their implementation within the Update-WorkItem's respective User switch actions

* Converting Invoke-WebRequest to Invoke-RestMethod (#94)

* Move WebRequest to RestMethod

Converted use of Invoke-WebRequest to Invoke-RestMethod for Cireson portal actions when Windows Authentication is not being used

* Convert API token into function

The PowerShell required to retrieve a Cireson API Token was moved into it's own dedicated function to when Windows Authentication is not used. Also caught an error with an extra "/" in the URL.

* Removing duplicate variables on Cireson calls

With the simplification of retrieving Cireson API tokens introduced, there are several Cireson based functions that have repetitive variables that could be condensed into a single one.

* Standardizing Cireson calls

This is more of a preference/consistency thing above all else in that all Cireson functions now declare their API URL specific variable before each Windows Auth if logic test. In this way they are more consistent and readable.

* Increase Get-SCSMUserByEmailAddress usage (#96)

* Leverage Get-SCSMUserByEmailAddress

Starting a branch to deploy a wider user of Get-SCSMUserByEmailAddress throughout New-WorkItem and Update-WorkItem to remove repetitive PowerShell.

* Updating CommentLeftBy logic

Condensing Update-WorkItem to use Get-SCSMUserByEmailAddress to decrease the lines of PowerShell needed to understand who left the comment/is performing the action on the Work Item being updated

* Continued deployment of Get-SCSMUserByEmailAddress

Adding Get-SCSMUserByEmailAddress usage into New-WorkItem, Update-WorkItem, and other areas seen throughout the script that perform a SCSM user lookup by email address

* Merge Get-SCSMAttachmentSettings and Get-SCSMObjectPrefix functions (#98)

* Merge Attachment Settings and Prefix functions

Merging Get-SCSMObjectPrefix into Get-SCSMAttachmentSettings to form Get-SCSMWorkItemSettings. This new function extends Get-SCSMAttachmentSettings to include work item prefix retrieval. In this regard, compatibility with it's original purpose is maintained. A few variables within Attach-FileToWorkItem and Attach-EmailToWorkItem have been renamed for context.

* Removing prefix requirement

Previously the first two characters were required to retrieve work item settings. With the new single function that does this by class name, this is no longer a requirement and helps move towards using custom prefixes.

* Matching case for "Prefix"

Just updating to match the case leveraged in the function

* Date values for keywords (#104)

Keywords that should set dates on Work Items don't. Also extending [take] keyword to mark FirstAssignedDate if it is not already set

* Simplify New-WorkItem by introducing Suggestion functions (#100)

* Condense New-WorkItem

Initial commit that removes some of the repetition seen in New-WorkItem when leveraging suggestions with/without Azure Cognitive Services. Creates two new functions Get-CiresonSuggestionURL and Send-CiresonSuggestionEmail that shorten New-WorkItem and simplify the logic when optionally leveraging Azure Cognitive Services

* Changing Send-CiresonSuggestionEmail

When Send-CiresonSuggestionEmail is called it was expecting an SCSM User Object, but when used it's missing the needed EmailAddress property as Email Addresses are technically a relationship in SCSM as opposed to a plain text field within Active Directory. Parameters have been changed, it's 6 references, and instead of passing the $affectedUser object instead am making use of the $from originally defined at the top of New-WorkItem

* MA/CR Support Group fix

Addressing #85 wherein -Class should have been used instead of -Classname. Also relocated section should these properties be set per Tom.

* Dropping Get-SCSMClassProperty for native property

Some inconsistences have arisen based on how/where Get-SCSMClassProperty is used. Rather than perform this "get" style operation this pivots into some of the more native SCSM C# SDK properties that are natively available via SMLets, drops the need for the cmd-let, and splatting.

* #84 Override WI Type by keyword match (not ACS) (#108)

* #84 Override WI Type by keyword match (not ACS)

* Re-arranged placement and expanded text

Relocated override functionality beneath Azure Cognitive Services section to logically group/keep the AI style features together.

* Keyword WI Type Override var name changed

Comments also updated and indicate that ACS takes priority over keyword override if both are enabled.

* #83 Consider BCC recipients in mailbox redirection (#107)

If no redirected mailboxes are found in the TO or CC fields, look in the Resent-From or Return-Path header, which will show that the address redirected mail to the main mailbox.  It can be safely inferred that the redirecting mailbox was a BCC recipient.

* Preparing for merge to master

Updating change log and changing how the config region is defined with the introduction of a matching closing section.