Introducing dynamic assignment
Based on the template that was applied for the New Work Item, optionally enable the ability to assign an analyst within that template's Support Group either entirely randomly or by the least amount of assigned Work Items. These two settings can be further enhanced by leveraging the SCSM Out of Office management pack that maintain their purpose but will overlook an Analyst whose Out of Office is true.