Excluding Work Item's that aren't being worked on
When determining the least amount of currently assigned work items for Dynamic Assignment, the counter of assigned work items per analyst should not count items that are in some form of a "finished" state such as Cancelled, Resolved, Skipped, Closed, etc. as it potentially skews fairness of assignment.

That said, this will count items that are "Open", includes Pending, and any custom sub-categories.