Dynamic Work Item Assignment (#111)
* Introducing dynamic assignment

Based on the template that was applied for the New Work Item, optionally enable the ability to assign an analyst within that template's Support Group either entirely randomly or by the least amount of assigned Work Items. These two settings can be further enhanced by leveraging the SCSM Out of Office management pack that maintain their purpose but will overlook an Analyst whose Out of Office is true.

* Excluding Work Item's that aren't being worked on

When determining the least amount of currently assigned work items for Dynamic Assignment, the counter of assigned work items per analyst should not count items that are in some form of a "finished" state such as Cancelled, Resolved, Skipped, Closed, etc. as it potentially skews fairness of assignment.

That said, this will count items that are "Open", includes Pending, and any custom sub-categories.

* Accidental variable drop

Between testing/commit removed the key means to actually count assignment

* Issue on pipeline

Breaking this into a few lines due to issue with selecting sourceobject

* Feature icon

Uploading icon for readme edit

* verbiage

initial verbiage around the assignment functionality

* Setting First Assigned Date

Given the property exists, it makes sense to set the First Assigned Date once the Work Item is dynamically assigned out through this feature