Updated Dynamic Assignment integration for AML
In order to get AML to work with the recently added Dynamic Work Item assignment feature, the function of Set-AssignedToPerTemplateSupportGroup was modified to take a SupportGroupID instead of a Template. In doing so, the feature can be used with or without AML and serves a more general purpose use. Finally, the function was renamed to Set-AssignedToPerSupportGroup