Merge reply & work item settings
Get-SCSMWorkItemSettings has now been modified to return the core of the prefix regex, so instead of returning the full match logic of "\[[I][R][0-9]+\]" for an Incident, it now simply returns [I][R]. As such, this requires the core matching logic of the connector to need a prefix and suffix of the square bracket and following number range. But enables a less (potentially) complex Merge Reply and Calendar regex. Finally, this does away with the C# style cast of [Regex]::Escape and returns to PowerShell.