Support for Custom Work Item Prefixes (#109)
* First draft

Beginning to address a long standing request for supporting custom Work Item prefixes. Still need to address building custom RE:/Reply regex for the Merge Replies functionality. Expanded Get-SCSMWorkItemSettings function to build the regex from the retrieved prefix.

* replacing hardcoded prefixes

Continuing to work on removing any hardcoded prefixes from the connector.

* Merge reply & work item settings

Get-SCSMWorkItemSettings has now been modified to return the core of the prefix regex, so instead of returning the full match logic of "\[[I][R][0-9]+\]" for an Incident, it now simply returns [I][R]. As such, this requires the core matching logic of the connector to need a prefix and suffix of the square bracket and following number range. But enables a less (potentially) complex Merge Reply and Calendar regex. Finally, this does away with the C# style cast of [Regex]::Escape and returns to PowerShell.

* Prefix/Suffix for New Regex logic

Updating the matching logic that was used when identifying a Closed Incident to get in line with the previous commit wherein Get-SCSMWorkItemSettings was modified to only return "[I][R]" for a prefix regex instead of "\[[I][R][0-9]+\]"

* IR w/ Activity Missing Prefix

Addressing bug that was identified during testing of custom prefixes. While it is it not the direct result of this new feature, it's something that can and should be addressed.

* Undoing patch

Looks like this has raised a merge conflict. Going to revert and deploy with a new branch.