More validation
Given the importance of keywords to the connector, I really wanted to ensure these incredibly core settings can't be entered incorrectly or accidentally. With the featured regex, keywords can't start or end with a space nor feature any special characters just like the PowerShell configuration. On Cireson portal integration, I opted to simply do a string check and append an "/" if it was missing. Finally, Problem Support Group has been opened up as a drop down.