Changing Class Extension logic
Previously, the way to retrieve valid class extensions was looking for a specific data type (string, dec, enum) and whose identifier did not contain X. As I'm reviewing this, #130 came to mind and as a result I've changed how the "not like" part of this statement works. Now this is based on the internal name of an MP that the property comes from. This should prove to be far more reliable.