Misleading verbiage
The checkbox controls whether or not autodiscover is used, but is phrased backwards. Since this flips the boolean logic for "isenabled" on the autodiscover text box, that's been removed and now it's always available for edit regardless of checkbox setting. Not ideal from a validation standpoint, but is less confusing.