Fixing Suggestion Call
Function doesn't handle when both values are null. As such, the function runs through it's entire execution only to error out. When returned to the New-WorkItem function, error are thrown to the effect of "Cannot index into a null array." It seems highly likely this isn't unique to v2. Will investigate master branch.