v1.6.1 (#165)
* Multi browser fix (#152)

Proposed fix for supporting different browsers and the Send Outlook email task in v1.x

* Refactoring take (#154)

Updating take logic so Tier Membership is only tested when required

* Improving MA actions (v1.x) (#156)

* Improving MA actions

Moving away from IF blocks in MA actions to more closely align with the other primary work item types (e.g. Change Request).

* updating MA take action

removed the take action when the email is from the implementer as it is redundant

* Suggestion fix (#161)

Updating the function that obtains suggestions to return null if the feature isn't configured. New-WorkItem is modified accordingly.

* syntax, update-workitem (#162)

syntax error was introduced on previous commits during #154

* fix, MA non-Implementer action (v1.x) (#163)

* fix, MA non-Implementer action

This is to preserve functionality in Update-WorkItem for Manual Activities when the comment to add is not from the Activity Implementer and features no keywords. The comment should be added to the MA Notes, rather than the Parent Work Item.

* changing user variable

the ActivityImplementer variable doesn't exist in the default context, shifting to CommentLeftBy

* updating SMLets link

had a specific version of smlets instead of the parent/most recent release link

* filename fix, again.

File was persisted through branching

* missing parameter name

Get-SCSMWorkItemParent was missing the named parameter

* updating version notes

updating version notes for next release

* notes (v1.6.1)

updating to include issue type