v2.0.1 (#166)
* Refactoring take

Updating take logic so Tier Membership is only tested when required

* Reverting take refactor

Didn't create a branch for it, restoring from previous version

* Multi browser fix (#153)

Proposed fix for supporting different browsers and the Send Outlook email task in v2.x

* Refactoring take (attempt 2) (#155)

Updating take logic so Tier Membership is only tested when required

* Improving MA actions (v2.x) (#157)

* Improving MA actions

Moving away from IF blocks in MA actions to more closely align with the other primary work item types (e.g. Change Request).

* updating MA take action

removed the take action when the email is from the implementer as it is redundant

* syntax error

previous commit broke switch logic via }

* Support Group extension fix (#159)

How the class extension for support group is found on MA, CR, and PR classes uses a different property name compared to v1

* fix, MA non-Implementer action (v2.x) (#164)

* fix, MA non-Implementer action

This is to preserve functionality in Update-WorkItem for Manual Activities when the comment to add is not from the Activity Implementer and features no keywords. The comment should be added to the MA Notes, rather than the Parent Work Item.

* changing user variable

the ActivityImplementer variable doesn't exist in the default context, shifting to CommentLeftBy

* updating SMLets link

had a specific version of smlets instead of the parent/most recent release link

* missing parameter name

Get-SCSMWorkItemParent was missing the named parameter

* updating version

changing version for next release

* updating notes

Updating notes for changes in versioning