watchlist validation logic
Introducing a new configuration variable that controls whether or not watchlist integration is used. Given the name of the variable I expanded on the notes to clear any potential confusion that could arise when changing it. This choice was intentional as v2.x of the connector abstracts this nuance away through the GUI.