Watchlist integration (v1.x) (#167)
* watchlist integration

Introducing the core add/remove functions for the watchlist feature along with the variables to define keywords

* add watchlist keyword actions

adds the watch/stopwatch feature to Incident, Service Request, Problem, and Change Request

* watchlist validation logic

Introducing a new configuration variable that controls whether or not watchlist integration is used. Given the name of the variable I expanded on the notes to clear any potential confusion that could arise when changing it. This choice was intentional as v2.x of the connector abstracts this nuance away through the GUI.

* correct watchlist variable

logic check on the "remove from watchlist" actions was testing the string length of addWatchlistKeyword instead of removeWatchlistKeyword