Watchlist integration (v2.x) (#168)
*  watchlist integration

Introducing the core add/remove functions for the watchlist feature along with the variables to define keywords

* add watchlist keyword actions

adds the watch/stopwatch feature to Incident, Service Request, Problem, and Change Request

* watchlist validation logic

Introducing a new configuration variable within the script that controls whether or not watchlist integration is used. This is already present in the MP as it controls whether or not the portal functionality is grayed out. Also updating keyword properties.

* watchlist keywords

extending the class to include the new keywords for adding and removing from the watchlist

* Watchlist data bindings

Updating AdminSettingsWizardData (C# code behind to the Settings UI) to reference the new Watchlist keyword properties on the Management pack (Settings.mpx)

* watchlist fields

adding watchlist fields for add/remove actions