v2.2.0 (#181)
* multi-mailbox logic

When the multi-mailbox feature isn't used PowerShell will throw errors around the feature. Rather than immediately test for the templates, this wraps a quick logic check into the New-WorkItem function.

* verify-workitem null check

Verify-WorkItem is being modified to handle the scenario where the file attachment search returns no results. In this case, a New Work item is created.

* Incorrect variable (#183)

Updating the final catch so to speak of the Get-TemplatesByMailbox function wherein the default workflow account's mailbox template is used.

* aml, predict config items (2.x) (#185)

* aml, predict config items

creating branch for extending the AML feature out to include the ability to predict impacted configuration items on IR/SR

* missing $

almost forgot to include the terminating $ in the Relationship Class name

* Min Confidence for Config Items

Adding new property for minimum impacted configuration item confidence

* updating save/load operations

updating the code behind to retrieve and set the value

* field for min confidence value

Adds a text field for entering the minimum confidence level

* AML Config Item icon

adding new aml icon with config item for new feature

* updating feature list

updating readme to include new AML feature that predicts impacted config items

* updating version number

changing the version number to be aligned with the next release

* changing next version

updating the next version so on build it matches the next release

Co-authored-by: TomH-OpsDev <35314798+TomH-OpsDev@users.noreply.github.com>