Azure Vision functions
Initial commit that declares the functions for Azure Vision Services that enable Azure to return words that describe the image (Get-AzureEmailImageAnalysis) and extract text from an image via OCR (Get-AzureEmailImageText). Will need to update both functions to work with binary streams instead of URLs so that as the connector parses attachments they can be uploaded and parsed by Azure within a processing loop of the connector.