Vision flow
Introduced primary config variable for Azure Vision along with a variable placeholder for the API Key. Modified Attach-FileToWorkItem function with conditions to check for the file type to then decide if to send off to Azure Vision. Still need to update said Azure functions will support for memory stream/byte array instead of a local image.