File extension parsing
When an attachment is retrieved through Exchange EWS, there does not appear to be a dedicated "file extension" property. As such, a small parsing mechanism has been introduced to Attach-FileToWorkItem so that the Extension property is now always filled out. This now makes the if statement that tests for common image types possible.