logic for analysis and OCR
Attach-FileToWorkItem now contains logic to decide how to define the Description property from Azure Vision. The top 5 image tags will always be written and if one of them is the word "text" then an OCR extraction attempt is made. If successful, it's appended to the Tags and then written into the File Attachment Description property