Azure Vision integration (v1.x) (#187)
* Azure Vision functions

Initial commit that declares the functions for Azure Vision Services that enable Azure to return words that describe the image (Get-AzureEmailImageAnalysis) and extract text from an image via OCR (Get-AzureEmailImageText). Will need to update both functions to work with binary streams instead of URLs so that as the connector parses attachments they can be uploaded and parsed by Azure within a processing loop of the connector.

* Vision flow

Introduced primary config variable for Azure Vision along with a variable placeholder for the API Key. Modified Attach-FileToWorkItem function with conditions to check for the file type to then decide if to send off to Azure Vision. Still need to update said Azure functions will support for memory stream/byte array instead of a local image.

* File extension parsing

When an attachment is retrieved through Exchange EWS, there does not appear to be a dedicated "file extension" property. As such, a small parsing mechanism has been introduced to Attach-FileToWorkItem so that the Extension property is now always filled out. This now makes the if statement that tests for common image types possible.

* Including try/catch for file extension

Wrapping previous statements in a try/catch just for the Extension property in the event a file is attached that doesn't end with a ".something"

* byte array instead of memorystream

Updating the call to the Azure Vision function to use the original Attachment Content's Byte Array of data as opposed to its Memory Stream.

* logic for analysis and OCR

Attach-FileToWorkItem now contains logic to decide how to define the Description property from Azure Vision. The top 5 image tags will always be written and if one of them is the word "text" then an OCR extraction attempt is made. If successful, it's appended to the Tags and then written into the File Attachment Description property

* url to byte array input

Azure Vision functions now take a byte array directly from the EWS Attachment

* updating readme

Adding information for Azure Vision integration