Workflow support (#208)
* SecureReference properties

Introducing properties to store the Id for SecureReferences for future lookup

* ComboBox control

resizing form to support and introducing Combo Box dropdown to support list of Run As Accounts

* workflow form

Adding dedicated form for controlling whether or not the workflow is enabled and how often it should run

* include workflow form

adding workflow form/resources to build

* workflow icon

adding workflow icon for the workflow form/wizard step

* workflow placeholder

Creating section for workflow engine feature

* workflow icon

including for readme

* screenshot alignment

forcing center alignment of graphic

* load secure reference bindings

Defining variables/bindings for loading a list and loading the selected item for the Exchange Web Services connection

* save secure reference id

saving the selected secure reference to the MP for easier lookup/load

* file path for connector

extending settings class with new data point to store the location of the smletsexchangeconnector.ps1 file

* connector text field

adding new text field for smlets exchange connector PowerShell file path

* browse button action

updating code behind for DLL xaml for the Browse button for finding the smlets exchange connector PowerShell script file path

* combobox control

resizing form to support and introducing Combo Box dropdown to support list of Run As Accounts

* cireson portal secure reference bindings

updating bindings between the UI and the MP for saving the ID of the secure reference to be used when authenticating to the cireson portal

* reference to new project

Adding build reference to new project within the solution for "SMLets.Exchange.Connector.Resources"

* workflow assembly project

New project to be included in the SMLets Exchange Connector solution to compile the DLL file to use the workflow

* workflow dependencies

With the introduction of the workflow, new dependencies are required to run the workflow/powershell

* update workflow

Adding logic that modifies an already existing instance of the SMLets Exchange Connector Workflow (i.e. Monitoring Rule). This updates if the workflow is enabled, how often the workflow runs per the UI and which secure reference to use

* updating wiki url

made the URL for configuring workflows ready for final release

* update workflow improvement

previously hard coded the management pack alias name. Now it's dynamically retrieved based on the Secure Reference that was selected and then used to update the Rule's XML

* create workflow - first pass

Rule creation still incomplete as it lacks a parent XML nodes for Data Source and Write Action that exist outside of the currently defined XML strings

* create workflow - second pass

The logic now exists to create a Workflow (monitoring) Rule from scratch if it doesn't exist. Thereby closing the loop between update and create processes.

* smexco file path binding

adds bindings to the settings UI for the smexco file path

* EWS variables

updating variables names for EWS to be more unique within the script

* About verbiage

updating the wording on the about page given the advent of this version's new functionality

* parameter case

updating case for parameters to match dependency registers of runscript.cs

* workflow ui bindings

bindings between the UI and setting the workflow values

* workflow wizard step

adding the workflow form to the wizard

* missing try/catch for portal run as

if the value isn't set the the MP would throw an error

* load workflow settings

load the workflow settings to the UI from the SCSM LFX mp if they are present

* workflow logic

the introduction of determining if the workflow is being used eliminates the need for renaming the ews variables