info - new-workitem
basic logging for the New-WorkItem function. Shows the message that came from Exchange and was passed into the function that's about to be used to create the work item, then SCSM specific concepts around From -> Affected User translation, and finally the created work item that was created. Given usage of ACS, AML, or regex patterns could be used to override the default work item type. This logging also highlights the initial work item type and potentially what the final result was