logging (#210)
* logging function

Still work to be done here (such as more validation sets per function and creating the event log), but this begins the standardization of logging regardless of deployment strategy.

* error - failure to read inbox

If the connection to Exchange can't be made, an event is recorded and the connector stops further processing

* info - messages processed

logging for the number of messages to be processed and then a counter that increments as each message is processed

* info - new-workitem

basic logging for the New-WorkItem function. Shows the message that came from Exchange and was passed into the function that's about to be used to create the work item, then SCSM specific concepts around From -> Affected User translation, and finally the created work item that was created. Given usage of ACS, AML, or regex patterns could be used to override the default work item type. This logging also highlights the initial work item type and potentially what the final result was

* info - update-workitem

introductory logging for update-workitem function

* info - update-workitem - RA

adding verbose logging support for review activities

* error - ews connection

moving failure to connect to exchange to a critical error which requires logging to be enabled

* logging - verify workitem

adding a mix of info and warning logging to Verify-WorkItem function

* logging icon

adding logging icon for readme

* logging settings

adding log properties

* logging bindings

creating bindings between ui and settings mp

* logging forms

adding forms for the logging page

* include logging forms

adding logging forms to the build

* logging wizardstep

adding a new wizard step to the wizard story for the new Logging page/form

* pull logging from settings

Creating variable to pull log settings from the Settings MP

* new-smexcoevent updates

making sure the function references the logging variables from the Settings MP and that the Event Log actually exists so the function can write to the Windows Event Log in the event the logging is enabled and it's being run from the workflow

* messages to process and variable expansion

changing logging so that messages to process must equal some value, lest generate event log noise when there is nothing to process. Also found variable expansion missing for CC/Related Users

* update-workitem, try/catch to if/else

moving the logging in update-workitem to an if/else to properly capture the difference between an Action and a Comment

* typo

correcting a mispelling